refactor: 마이대시보드 페이지, 사이드바 로직 hook으로 정리
