+++ components/organisms/SideBar/SideBar.tsx
@@ -1,13 +1,9 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-
-import { useQuery } from '@tanstack/react-query';
 import Image from 'next/image';
 import Link from 'next/link';
 import styled from 'styled-components';
 
-import { Dashboard, getDashboardList } from '@apis/dashboards/getDashboardList';
 import createDashboardIcon from '@public/images/icons/add-filledGray_787486-w20-w20.svg';
 import crown from '@public/images/icons/crown-filledYellow-FDD446-w16-h12.svg';
 import LogoSvg from '@public/images/logos/logo-small-filledViolet-w28.82-h33.07.svg?component';
@@ -15,48 +11,19 @@ import TaskifySvg from '@public/images/logos/taskify-text-small-filledViolet-w80
 import { mediaBreakpoint } from '@styles/mediaBreakpoint';
 
 import ColoredDot from '@components/atoms/ColoredDot';
-import { SIDE_BAR_PAGE_GROUP_NUMBER } from '@components/constants';
 
-import useModal from '@hooks/use-modal';
-
+import useSideBar from './hook';
 import PageNationButton from '../../atoms/PageNationButton';
 
 export default function SideBar() {
-  const [currentPage, setCurrentPage] = useState<number>(1);
-  const [totalPage, setTotalPage] = useState<number>();
-
-  const { data, isSuccess } = useQuery({
-    queryKey: ['dashboard', 'dashboardList', currentPage, SIDE_BAR_PAGE_GROUP_NUMBER],
-    queryFn: () => getDashboardList(currentPage, SIDE_BAR_PAGE_GROUP_NUMBER),
-  });
-  const [dashboards, setDashboards] = useState<Dashboard[]>(data?.dashboards || []);
-
-  const handleNextDashboardPageClick = () => {
-    setCurrentPage((prev) => prev + 1);
-  };
-
-  const handlePreviousDashboardPageClick = () => {
-    setCurrentPage((prev) => prev - 1);
-  };
-
-  // 대시보드 리스트 조회
-  useEffect(() => {
-    if (isSuccess) {
-      setDashboards(data?.dashboards);
-      setTotalPage(Math.ceil(data.totalCount / SIDE_BAR_PAGE_GROUP_NUMBER));
-    }
-  }, [data, isSuccess, currentPage]);
-
-  // 대시보드 생성 버튼 클릭
-  const { openModal } = useModal();
-
-  const handleCreateDashboardButtonClick = async () => {
-    const CreateDashboardModal = await import('@components/organisms/CreateDashboardModal').then(
-      (module) => module.default,
-    );
-
-    openModal(CreateDashboardModal);
-  };
+  const {
+    dashboards,
+    totalPage,
+    currentPage,
+    handleNextDashboardPageClick,
+    handlePreviousDashboardPageClick,
+    handleCreateDashboardButtonClick,
+  } = useSideBar();
 
   return (
     <S.Wrapper>
